requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv
aiohttp
aiofiles
structlog
prometheus-client
pytest
pytest-asyncio
